refactor: app.ts 수정
